requirements.txt
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0